backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
PyPDF2
spacy
scikit-learn
numpy

